Standardizing module organization. Moving safe apis to <module>/safe.rs
