+++ src/cublas/mod.rs
@@ -1,3 +1,6 @@
 pub mod result;
+pub mod safe;
 #[allow(warnings)]
 pub mod sys;
+
+pub use safe::*;
