+++ src/curand/mod.rs
@@ -1,7 +1,9 @@
 //! Wrappers around the [cuRAND API](https://docs.nvidia.com/cuda/curand/index.html)
-//! in two levels: [result], and [sys].
-//! See [crate::rng] for safe wrappers.
+//! in three levels. See crate documentation for description of each.
 
 pub mod result;
+pub mod safe;
 #[allow(warnings)]
 pub mod sys;
+
+pub use safe::*;
