+++ src/driver/mod.rs
@@ -1,9 +1,9 @@
 //! Wrappers around the [CUDA driver API](https://docs.nvidia.com/cuda/cuda-driver-api/index.html),
-//! in two levels: [result], [sys]. See [crate::device] for safe wrappers.
-//!
-//! 1. [result] thin wrapper around [sys] to return [Result<_, result::DriverError>]
-//! 2. [sys] the raw bindings
+//! in three levels. See crate documentation for description of each.
 
 pub mod result;
+pub mod safe;
 #[allow(warnings)]
 pub mod sys;
+
+pub use safe::*;
